backend/becompanyportal/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Django>=5.0,<6.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
executing==2.0.1
graphviz==0.20.3
inflection==0.5.1
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
setuptools==71.1.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
uritemplate==4.1.1
wcwidth==0.2.13
